requirements.txt
tensorflow==2.4.0
matplotlib==3.2.2
numpy==1.18.5
opencv-python==4.1.2
Pillow==8.0.0
PyYAML==5.3.1
scipy==1.4.1
torch==1.7.0
torchvision==0.8.1
tqdm==4.41.0
arabicocr==1.1.6
arabic_reshaper==2.1.3
easyocr==1.4.1
h5py==2.10.0
pandas==1.3.3
pytorch==1.9.1